make fact checking index page
